Moved LootItemData to it's own file to avoid cross importing. Started working on frontend of loot log
