+++ templates/loot.html
@@ -1,0 +1,6 @@
+{% extends base.html %}
+
+{% block content %}
+
+
+{% endblock %}
